Add Code To Find Black Scenes/Scene Changes
Wrote a script that takes a video as input, and uses ffmpeg to find
black scenes (saves the start time and length in seconds). Also prints
code that can be run in the command line to find scene changes (unable
to figure out syntax to get it to work in Python subprocess).